[Sentry] Move breadcrumb definitions off main SentryHelper.cs
